+++ tests/doc.rs
@@ -3,8 +3,7 @@
 //! ```
 use std::io;
 
-use tap::Pipe;
-use testutils::os_cmd::{Runner, presets::CargoDoc};
+use testutils::os_cmd::{RunnableCommand, presets::CargoDoc};
 
 #[ignore]
 #[test]
@@ -15,7 +14,6 @@ fn build_and_open_rust_doc() -> io::Result<()> {
   ] {
     CargoDoc::default()
       .with_pkg(pkg)
-      .pipe(Runner::from)
       .run()?
   }
   Ok(())